ticket-generation-and-bulkmailer-python/requirements.txt
Pillow
requests
qrcode
python-dotenv
